22-07-2026 | Upgraded UI for View Claim Fragment and background of other fragments

+++ app/src/main/java/app/xedigital/ai/model/TeamMember/ChildrenItem.java
@@ -40,7 +40,7 @@ public class ChildrenItem {
     private String createdAt;
 
     @SerializedName("ctc")
-    private int ctc;
+    private double ctc;
 
     @SerializedName("children")
     private List<Object> children;
@@ -179,7 +179,7 @@ public class ChildrenItem {
         return createdAt;
     }
 
-    public int getCtc() {
+    public double getCtc() {
         return ctc;
     }
